add google verification code and fix views counter
