+++ src/pages/api/v1/views.tsx
@@ -52,7 +52,7 @@ const registerView = async (req: NextApiRequest, res: NextApiResponse) => {
 
     res.status(200).json({ views: views.count, success: true });
   } catch (error) {
-    console.log(error);
+    // console.log(error);
     res.status(500).json({ error });
   }
 };
